id ve version olan heryer Long yerine Integer oldu

+++ src/main/java/com/aop/IUserActivityDao.java
@@ -5,6 +5,6 @@ import com.entity.UserActivity;
 
 public interface IUserActivityDao extends IBaseDao<UserActivity>{
 	
-	public Long addUserActivity(UserActivity userActivity);
+	public Integer addUserActivity(UserActivity userActivity);
 	
 }
