+++ src/main/java/com/controller/EmployeeController.java
@@ -1,15 +1,16 @@
 package com.controller;import java.io.Serializable;
 import java.util.ArrayList;
+
 import javax.annotation.PostConstruct;
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ManagedProperty;
 import javax.faces.bean.ViewScoped;
 import javax.faces.context.FacesContext;
+
 import org.apache.log4j.Logger;
+
 import com.dao.IEmployeeService;
 import com.entity.Employee;
-import com.enums.ProjectEnum.RelationType;
-import com.enums.ProjectEnum.AddSelect;
 
 @ViewScoped
 @ManagedBean(name = "employeeController")
@@ -37,7 +38,7 @@ getEntityList().clear();
 }
 
 @Override
-public void setEid(Long eid) {
+public void setEid(Integer eid) {
 setEntity(employeeService.getEntityById(eid));
 this.eid = eid;
 }
