+++ src/main/java/com/controller/PaymentController.java
@@ -1,11 +1,14 @@
 package com.controller;import java.io.Serializable;
 import java.util.ArrayList;
+
 import javax.annotation.PostConstruct;
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ManagedProperty;
 import javax.faces.bean.ViewScoped;
 import javax.faces.context.FacesContext;
+
 import org.apache.log4j.Logger;
+
 import com.dao.IPaymentService;
 import com.entity.Payment;
 import com.enums.ProjectEnum.RelationType;
@@ -36,7 +39,7 @@ getEntityList().clear();
 }
 
 @Override
-public void setEid(Long eid) {
+public void setEid(Integer eid) {
 setEntity(paymentService.getEntityById(eid));
 this.eid = eid;
 }
